Main algorithm
INPUT:

Launch code and enter digits found in each horizontal line in the grid of the Xsudoku puzzle one by one, with empty squares represented by zero. For example 010258003. Then, press "Enter" and enter the next line of the puzzle.

OUTPUT:

Two grids, where the upper grid demonstrates the puzzle and the lower grid demonstrates the solution. In between them is the number of empty squares that the algorithm will have to solve for and the solution process. After the lower grid with the solution, there is the execution time and the number of plotted digits (must be equal to the number of empty squares).